+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changes
 
 * Add ``subsystems`` definitions to blocks stored in the blockchains to detail referenced data for which the
   consents are being applied.
+* Add more extensive documentation about content metadata when ``subsystems`` are involved.
 * Fix import error when computing block hash.
 
 `1.1.0 <https://github.com/crim-ca/blockchain/tree/1.1.0>`_ (2022-01-26)